Dockerise auto_tesseract (#1)
This package requires a few unusual system dependencies (e.g; tesseract) and some special configuration for ImageMagick. It makes more sense to isolate these changes to a container, rather than as a system-wide package.

To keep matters simple, the Docker image contains the whole kit-and-kaboodle: app, tests, and test/lint libraries/binaries and the process driven from a Makefile (rather than bazel). The github workflow therefore needs to be redone, so it's removed for now.